changing size of n from int to long long

+++ app/app.cpp
@@ -67,7 +67,7 @@ static void doWork(int argc, char **argv)
   PrimeCheckerParams& params = parseCommandLineOptions(argc, argv);
   
   PrimeChecker primeChecker(params.verboseMode, params.runMode);
-  int n;
+  long long n;
 
   while(1)
     {
